Implement singleton pattern for APScheduler() class
Currently, when Flask-APScheduler is used in a WSGI
app, for instance, powered by Gunicorn, the APScheduler
class is instantiated for each worker process spawned by
the WSGI web-server (i.e. if I set the '--workers 4' flag
in gunicorn, APScheduler will get instantiated for each worker).

This is also an issue when we set DEBUG=True in the flask
app's config, as this will spawn a Master and a Child process.

The reason this is a problem is that all "n" instantiations of
APScheduler will collectively execute the job "n-times", in
essence doubling, tripling, quadrupling, etc. the execution of
each job.

If I start Gunicorn with 4 workers, and have a job that
dumps a MySQL database ever hour, this job will execute
4 times every hour, at the same time. This is obviously not
what we want. (Not to mention we have more memory overhead
if we have "n" instances of the APScheduler class)

I have provided a Singleton implementation of APScheduler,
which works as follows: If we try and create a new APScheduler
on a worker process, if a pre-existing APScheduler instance exists,
return it! This way, every worker process shares the Singleton
APScheduler object, and each job is executed only once, properly.